#69 Feat: 상세리뷰 스크롤 구현
